Split into separate files. .gitignore requirements.txt
